requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
boltons
boto3==1.28.65
botocore==1.31.65
brotlipy==0.7.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi
cffi
charset-normalizer
clearml==1.12.2
cleo==2.1.0
colorama
comm==0.1.4
contourpy==1.1.0
crashtest==0.4.1
cryptography
cycler==0.11.0
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.7
dulwich==0.21.6
executing==2.0.1
fastjsonschema==2.18.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
future==0.18.3
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.18.0
idna
imageio==2.31.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
installer==0.7.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch
jsonpointer==2.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.14.0
keyring==24.3.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
libmambapy
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
menuinst
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
orderedmultidict==1.0.1
overrides==7.4.0
packaging
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
proglog==0.1.10
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat
pycparser
Pygments==2.16.1
PyJWT==2.4.0
pyOpenSSL
pyparsing==3.0.9
pyproject_hooks==1.0.0
PySocks
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytube==15.0.0
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
rapidfuzz==3.5.2
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
rsa==4.9
Rtree==1.0.1
ruamel.yaml
s3transfer==0.7.0
scikit-image==0.21.0
scikit-learn==1.3.2
scikit-video==1.1.11
scipy==1.11.1
seaborn==0.13.0
Send2Trash==1.8.2
shapely==2.0.1
shellingham==1.5.4
six
sk-video==1.1.10
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
super-image==0.1.7
sympy==1.12
telebot==0.0.5
tensorflow==2.14.0
torch=2.0.1
mediapipe